Added --dump-all
git-svn-id: svn://svn.cc65.org/cc65/trunk@241 b7a2c559-68d2-44c3-8de9-860c34a00d81

+++ src/od65/global.c
@@ -43,7 +43,7 @@
 
 
 
-unsigned long	What	= 0;		/* What should get dumped? */
+unsigned What	= 0;		/* What should get dumped? */
